// motor-insurance-frontend/src/types/client.ts
export interface ClientPolicy {
  policyId: string;
  registrationNumber: string;
  make: string;
  model: string;
  colour: string;
  year: number;
  vehicleType: number;
  insuranceType: number;
  startDate: string;
  endDate: string;
  amount: number;
  status: string;
  durationQuarters: number;
}